meson.build: Set dep_x11 and dep_xext for panfrost
Using cc.find_library() for non-C stdlib dependencies only works if
you're guaranteed to have all of the dependencies you need installed
system wide. This isn't the case for cross compilation, which means that
we need to do things the right way so that the correct flags for linking
get specified during cross compilation. So, do that.

This, with the addition of

https://github.com/mesonbuild/meson/pull/4473

Fix cross compilation of panfrost for my aarch64 system