gallium/panfrost: Kill optional memory mapping
Nowadays mali_kbase tells us when we need to mmap() things, not the
other way around. Unfortunately it's a little rude and doesn't seem to
specify the right output flag for requiring an mmap, so we just tell if
an mmap is required by checking for the magic number that's always the
first cookie. Let's hope we never try multithreading with this!

This might have something to do with the kbase API we're reporting.