panfrost: Support float/vec2/vec3 attributes
Previously, we had hard-coded the number of components to equal 4
(vec4), due to mysterious issues occurring when the count was less than
4. Further investigation shows that not only the component count needs
updating but also the magic value in the unknown1 field. That field
still needs to be further decoded - I suspect it controls how extra
components are padded out - but for now, this patch allows these
attributes to work at all.

This enables Sway to render.